requirements.txt
PyMuPDF==1.23.8
pdfplumber==0.10.3
PyPDF2==3.0.1
pdfminer.six==20221105 